+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+* [#44](https://vscode.dev/github/sdss/lvmgort/pull/44) RORR RID-019: disables the Overwatcher if rain is detected and requires a human to re-enable it when conditions are safe.
+
+
 ## 1.3.0 - November 29, 2024
 
 ### 🚀 New