+++ CHANGELOG.md
@@ -7,6 +7,15 @@
 * [#44](https://vscode.dev/github/sdss/lvmgort/pull/44) RORR RID-019: disables the Overwatcher if rain is detected and requires a human to re-enable it when conditions are safe.
 * Create the night log during the pre-observing task.
 
+### 🏷️ Changed
+
+* Removed the `pubsub` module and use `lvmopstools` instead.
+
+### 🔧 Fixed
+
+* Do not run the post-exposure checks when cancelling the loop.
+* Fix call to `/notifications/create` for new API version.
+
 
 ## 1.3.0 - November 29, 2024